requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-openai
langchain-postgres
langgraph
neo4j
pandas
plotly
networkx
pypdf
psycopg2-binary
tiktoken
langchain-experimental
langchain-google-genai